requirements.txt
pandas~=1.2.3
librosa~=0.8.0
joblib~=1.0.1
tqdm~=4.56.0
scikit-learn~=0.24.1
numpy~=1.19.2
pytorch~=1.8.0
wandb~=0.10.26
torchvision~=0.9.0